requirements.txt
absl-py==2.0.0
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.13.1
Flask==3.0.0
flatbuffers==23.5.26
fsspec==2023.12.1
gast==0.5.4
google-auth==2.25.1
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.3
h5py==3.10.0
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
nltk==3.8.1
numpy==1.24.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
praw==7.7.1
prawcore==2.4.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
tensorflow==2.15.0
termcolor==2.4.0
textblob==0.17.1
threadpoolctl==3.2.0
tokenizers==0.15.0
tqdm==4.66.1
transformers==4.35.2
typing_extensions==4.8.0
update-checker==0.18.0
urllib3==2.1.0
websocket-client==1.7.0
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
